vcu_project/requirements.txt
python-can==4.3.1
adafruit-circuitpython-ads1x15==2.3.10
adafruit-blinka==8.27.0
RPi.GPIO==0.7.1
